Fix FileSystemSigner with PKCS#12 file
* Use *rsa.PrivateKey and *ec.PrivateKey types across the project to
  maintain consistency
* Refactor Makefile so that it's easy to run the subset of tests that
  don't require non-trivial dependencies
